refactor(apidocs): Tidy up annotations
