--- keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/docs/Literal.kt
@@ -1,6 +1,0 @@
-package com.netflix.spinnaker.keel.api.docs
-
-import kotlin.annotation.AnnotationTarget.CLASS
-
-@Target(CLASS)
-annotation class Literal(val type: String = "string", val value: String)
