--- keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/docs/Optional.kt
@@ -1,6 +1,0 @@
-package com.netflix.spinnaker.keel.api.docs
-
-import kotlin.annotation.AnnotationTarget.PROPERTY
-
-@Target(PROPERTY)
-annotation class Optional
